configured for session_id

+++ aimongus_types/src/client_to_server.rs
@@ -8,7 +8,7 @@ use ts_rs::TS;
 pub enum ClientResponse {
     StartGame,
     SubmitAnswer(String),
-    SubmitVote(u8),
+    SubmitVote(u128),
     ReadyForNextTurn,
 }
